+++ a2/cpanchag-hsavla-mmasrani-ritekavu/breakout_assgn_2_v_1_0/PaddleRightMoveCommand.java
@@ -2,18 +2,18 @@ package breakout_assgn_2_v_1_0;
 
 public class PaddleRightMoveCommand implements Command {
 
-	private int prevX;
+	private double prevX;
 	private Paddle paddle;
 
 	PaddleRightMoveCommand(Paddle paddle) {
 		this.paddle = paddle;
 	}
 
-	public int getPrevX() {
+	public double getPrevX() {
 		return prevX;
 	}
 
-	public void setPrevX(int prevX) {
+	public void setPrevX(double prevX) {
 		this.prevX = prevX;
 	}
